requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flasgger==0.9.7.1
python-dotenv==1.0.0
PyMySQL==1.1.0
psycopg2-binary==2.9.7
Werkzeug==2.3.7
gunicorn==21.2.0 